add condition to check environment of app

+++ src/prisma/prisma.service.ts
@@ -5,10 +5,13 @@ import { EnvConfigType } from '../config';
 
 @Injectable()
 export class PrismaService extends PrismaClient {
-  constructor(config: ConfigService<EnvConfigType>) {
+  constructor(private config: ConfigService<EnvConfigType, true>) {
+    const url = config.get<string>('DATABASE_URL');
     super({
       datasources: {
-        db: { url: config.get('DATABASE_URL') },
+        db: {
+          url,
+        },
       },
     });
   }
@@ -17,6 +20,14 @@ export class PrismaService extends PrismaClient {
    * @returns
    */
   cleanDb() {
+    //* agar muncul type completion pada saat pengecekan environment/lingkungan dimana aplikasi berjalan.
+    //* inisialisasi terlebih dahulu ke sebuah variabel
+    const env = this.config.get<EnvConfigType['NODE_ENV']>('NODE_ENV');
+    //* sehingga dapat meminimalisir kesalahan pengetikan.
+    //* pengecekan ini dilakukan untuk mencegah database terhapus secara tidak sengaja pada production environment.
+    //* atau juga bisa menggunakan type casting jika env tidak diberi generic type secara eksplisit
+    //* if ((env as EnvConfigType['NODE_ENV']) === 'production') return;
+    if (env === 'production') return;
     return this.$transaction([this.user.deleteMany(), this.note.deleteMany()]);
   }
 }
